requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
marshmallow==3.20.2
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.1
PyJWT==2.8.0
SQLAlchemy==2.0.27
Werkzeug==3.0.1
flask-cors==4.0.0
pytest==8.0.2
pytest-cov==4.1.0
pandas==2.2.1
openpyxl==3.1.2
ipython==8.22.2